+++ service/service.go
@@ -2,7 +2,6 @@ package service
 
 import (
 	"context"
-	"strings"
 	"sync"
 
 	applicationv1alpha1 "github.com/giantswarm/apiextensions/v2/pkg/apis/application/v1alpha1"
@@ -15,11 +14,9 @@ import (
 	"github.com/spf13/afero"
 	"github.com/spf13/viper"
 	"k8s.io/client-go/rest"
-	"sigs.k8s.io/yaml"
 
 	"github.com/giantswarm/app-operator/v2/flag"
 	"github.com/giantswarm/app-operator/v2/pkg/project"
-	"github.com/giantswarm/app-operator/v2/service/collector"
 	"github.com/giantswarm/app-operator/v2/service/controller/app"
 	"github.com/giantswarm/app-operator/v2/service/controller/appcatalog"
 )
@@ -40,7 +37,6 @@ type Service struct {
 	appController        *app.App
 	appCatalogController *appcatalog.AppCatalog
 	bootOnce             sync.Once
-	operatorCollector    *collector.Set
 }
 
 // New creates a new service with given configuration.
@@ -128,36 +124,6 @@ func New(config Config) (*Service, error) {
 		}
 	}
 
-	// TODO: Remove once appcatalogentry CRs exist and have the name of the
-	// team responsible for the app.
-	//
-	//	https://github.com/giantswarm/roadmap/issues/26
-	//
-	var appTeamMapping map[string]string
-	{
-		appTeamMapping, err = newAppTeamMapping(config.Viper.GetString(config.Flag.Service.Collector.Apps.Teams))
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
-	var operatorCollector *collector.Set
-	{
-		c := collector.SetConfig{
-			K8sClient: k8sClient,
-			Logger:    config.Logger,
-
-			AppTeamMapping: appTeamMapping,
-			DefaultTeam:    config.Viper.GetString(config.Flag.Service.Collector.Apps.DefaultTeam),
-			UniqueApp:      config.Viper.GetBool(config.Flag.Service.App.Unique),
-		}
-
-		operatorCollector, err = collector.NewSet(c)
-		if err != nil {
-			return nil, microerror.Mask(err)
-		}
-	}
-
 	var versionService *version.Service
 	{
 		versionConfig := version.Config{
@@ -181,7 +147,6 @@ func New(config Config) (*Service, error) {
 		appController:        appController,
 		appCatalogController: appCatalogController,
 		bootOnce:             sync.Once{},
-		operatorCollector:    operatorCollector,
 	}
 
 	return newService, nil
@@ -190,29 +155,8 @@ func New(config Config) (*Service, error) {
 // Boot starts top level service implementation.
 func (s *Service) Boot(ctx context.Context) {
 	s.bootOnce.Do(func() {
-		go s.operatorCollector.Boot(ctx) // nolint:errcheck
-
 		// Start the controllers.
 		go s.appCatalogController.Boot(ctx)
 		go s.appController.Boot(ctx)
 	})
-}
-
-func newAppTeamMapping(input string) (map[string]string, error) {
-	appTeamMapping := make(map[string]string)
-
-	teams := map[string]string{}
-
-	err := yaml.Unmarshal([]byte(input), &teams)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	for team, apps := range teams {
-		for _, app := range strings.Split(apps, ",") {
-			appTeamMapping[app] = team
-		}
-	}
-
-	return appTeamMapping, nil
 }